pyproject.toml
[project]
name = "menu"
version = "0.1.0"
description = ""
authors = [{name = "Mohit Nalavadi"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "responses>=0.24.0",
    "black",
    "jupyter>=1.1.1",
    "notebook>=7.5.1",
    "pandas>=2.3.3",
    "ruff>=0.14.10",
    "isort>=7.0.0",
    "typer>=0.9.0",
    "flask>=3.0.0",
    "openai>=1.12.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "joblib>=1.3.0",
    "vulture>=2.14",
    "pydantic>=2.0.0",
    "google-search-results>=2.0.0",
]

[tool.config]
flask_port = 5011
max_upload_size_mb = 10
default_target_currency = "EUR"
default_openai_model = "gpt-5-mini"

[project.scripts]
app = "src.app:main"
config = "src.config:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 110
target-version = ['py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.isort]
force_single_line = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-s -v --cov=. --cov-report=term-missing"

[tool.coverage.run]
omit = ["tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
